refactor: extract BottomButtons.tsx
